+++ src/watchdog/observers/inotify.py
@@ -123,6 +123,14 @@ class InotifyEmitter(EventEmitter):
             return path
         return unicode_paths.decode(path)
 
+    def run(self):
+        self._inotify.start()
+        return EventEmitter.run(self)
+
+    @property
+    def ready(self):
+        return self._inotify.ready
+
     def on_thread_stop(self):
         self._inotify.close()