+++ changelog.rst
@@ -3,6 +3,14 @@
 API changes
 -----------
 
+
+0.8.1 - unreleased
+~~~~~~~~~~~~~~~~~~
+
+- ``EventEmitter.start()`` no waits for emitter to start.
+- Don't start the emitter unless the main observer is not started.
+
+
 0.8.0
 ~~~~~